// src/app/model/tela-casaOracoes/casaOracoes.ts
export interface CasaOracoes {
  igrId?: number;
  igrCod?: string;
  igrNome?: string;
  admNome?: string; // Incluindo admNome
  setorNome?: string;
  igrEstado?: string;
  igrCidade?: string;
  igrBairro?: string;
  igrCep?: string;
  igrEndereco?: string;
  igrComplemento?: string;
  admId?: number;
  setorId?: number;
}